+++ civis_jupyter_notebooks/__main__.py
@@ -16,6 +16,8 @@ def cli():
     # make home areas and dirs
     for dr in [('~', 'work'),
                ('~', '.jupyter', 'custom'),
+               # folder that holds all the JS for notebook frontend extensions
+               ('~', '.jupyter', 'extensions'),
                ('~', '.jupyter', 'custom', 'fonts'),
                ('~', '.ipython', 'profile_default')]:
         try:
@@ -42,3 +44,10 @@ def cli():
     _copy(('assets', '.bashrc'), ('~'))
     _copy(('assets', 'ipython_config.py'), ('~', '.ipython', 'profile_default'))
     _copy(('assets', 'civis_client_config.py'), ('~', '.ipython'))
+
+    _copy(('assets', 'extensions', 'uncommitted_changes.js'), ('~', '.jupyter', 'extensions'))
+
+    # install and enable nbextensions
+    for cmd in ['jupyter nbextension install ~/.jupyter/extensions',
+                'jupyter nbextension enable extensions/uncommitted_changes']:
+        subprocess.check_call(cmd, shell=True)