+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.3.0] - 2017-12-12
 
+### Added
 - Improvements to terminal navigation
 - Default terminal to bash
 - Add civicon font
+- Git integration with notebooks
 
 ## [0.2.4] - 2017-11-28